requirements.txt
streamlit
groq
chromadb
sentence-transformers
PyPDF2
python-docx
langchain-text-splitters